// src/app/model/ville.ts
export class Ville {
    Ville: string;
    nom: string;
    insee: string;
    habitants: string;
    deliberation: string;
    FPS_mini: string;
    mini_minore: string;
    fps_mini_delai: string;
    FPS_maxi: string;
    maxi_minore: string;
    fps_maxi_delai: string;
    source: string;
    commentaire: string;
    codeDepartement: string;
    nomDepartement: string;
    populationMunicipale: string;
    populationTotale: string;
}